CRUD enabled

+++ src/main/java/com/dev/tugabesar/Controller.java
@@ -31,7 +31,7 @@ public class Controller {
     return false;
   }
 
-  public static boolean addData(String id, String nameItem, String category, String user, JFrame frame) {
+  public static boolean addData(String id, String nameItem, String category, String privilege, JFrame frame) {
     try {
       Connection myCon = Mysql.getConnection();
       String query = "INSERT INTO data (id, nameItem, category, user)" + " values (?, ?, ?, ?)";
@@ -41,7 +41,7 @@ public class Controller {
       preparedStatement.setString(1, id);
       preparedStatement.setString(2, nameItem);
       preparedStatement.setString(3, category);
-      preparedStatement.setString(4, user);
+      preparedStatement.setString(4, privilege);
 
       preparedStatement.execute();
       preparedStatement.close();
@@ -57,17 +57,72 @@ public class Controller {
   public static boolean deleteData(String id, JFrame frame) throws Exception {
     try {
       Connection myCon = Mysql.getConnection();
-      String query = "DELETE FROM data WHERE id = ?";
+      String query = "SELECT user FROM data WHERE id = ?";
       PreparedStatement preparedStatement = myCon.prepareStatement(query);
-
       preparedStatement.setString(1, id);
-      preparedStatement.executeUpdate();
-      preparedStatement.close();
 
-      return true;
-    } catch (SQLException error) {
-      JOptionPane.showMessageDialog(frame, error.getMessage());
+      ResultSet resultSet = preparedStatement.executeQuery();
+      while (resultSet.next()) {
+        Sesi.whoAreYou = resultSet.getString("user");
+      }
+
+    } catch (Exception error) {
+      System.out.println(error.getMessage());
     }
+
+    if (Sesi.whoAreYou.equals(Sesi.Privilege) || Sesi.Privilege.equals("admin")) {
+      try {
+        Connection myCon = Mysql.getConnection();
+        String query = "DELETE FROM data WHERE id = ?";
+        PreparedStatement preparedStatement = myCon.prepareStatement(query);
+
+        preparedStatement.setString(1, id);
+        preparedStatement.executeUpdate();
+        preparedStatement.close();
+
+        JOptionPane.showMessageDialog(frame, "Data terhapus");
+        return true;
+      } catch (SQLException error) {
+        System.out.println(error.getMessage());
+      }
+    }
+    JOptionPane.showMessageDialog(frame, "Anda tidak memiliki hak akses!");
+    return false;
+  }
+
+  public static boolean updateData(String id, String nameItem, String category, JFrame frame) {
+    try {
+      Connection myCon = Mysql.getConnection();
+      String query = "SELECT user FROM data WHERE id = ?";
+      PreparedStatement preparedStatement = myCon.prepareStatement(query);
+      preparedStatement.setString(1, id);
+
+      ResultSet resultSet = preparedStatement.executeQuery();
+      while (resultSet.next()) {
+        Sesi.whoAreYou = resultSet.getString("user");
+      }
+
+    } catch (Exception error) {
+      System.out.println(error.getMessage());
+    }
+
+    if (Sesi.Privilege.equals(Sesi.whoAreYou) || Sesi.Privilege.equals("admin")) {
+      try {
+        Connection myCon = Mysql.getConnection();
+        String query = "UPDATE data SET nameItem = ?, category = ? WHERE id=?";
+        PreparedStatement preparedStatement = myCon.prepareStatement(query);
+        preparedStatement.setString(1, nameItem);
+        preparedStatement.setString(2, category);
+        preparedStatement.setString(3, id);
+
+        preparedStatement.execute();
+
+        return true;
+      } catch (Exception error) {
+        System.out.println(error.getMessage());
+      }
+    }
+    JOptionPane.showMessageDialog(frame, "Anda tidak memiliki hak akses!");
     return false;
   }
